Revert "Add option to follow system theme & implement dark mode everywhere"
