+++ Alkitab/src/main/AndroidManifest.xml
@@ -43,7 +43,6 @@
 
 		<activity
 			android:name="yuku.alkitab.base.IsiActivity"
-			android:configChanges="colorMode|uiMode"
 			android:exported="true">
 			<intent-filter>
 				<action android:name="android.intent.action.MAIN" />
@@ -249,8 +248,7 @@
 			android:windowSoftInputMode="adjustResize" />
 		<activity
 			android:name="yuku.alkitab.base.ac.DevotionActivity"
-			android:label="@string/menuDevotion"
-			android:configChanges="colorMode|uiMode">
+			android:label="@string/menuDevotion">
 			<meta-data
 				android:name="android.support.PARENT_ACTIVITY"
 				android:value="yuku.alkitab.base.IsiActivity" />
@@ -277,8 +275,7 @@
 			android:windowSoftInputMode="adjustResize|stateHidden" />
 		<activity
 			android:name="yuku.alkitab.songs.SongViewActivity"
-			android:label="@string/sn_songs_activity_title"
-			android:configChanges="colorMode|uiMode" />
+			android:label="@string/sn_songs_activity_title" />
 		<activity
 			android:name="yuku.alkitab.base.ac.SecretSettingsActivity"
 			android:label="Secret settings" />
@@ -288,8 +285,7 @@
 		<activity android:name="yuku.alkitab.base.ac.ReadingPlanActivity" />
 		<activity
 			android:name="yuku.alkitab.base.ac.PatchTextActivity"
-			android:windowSoftInputMode="adjustResize"
-			android:configChanges="colorMode|uiMode" />
+			android:windowSoftInputMode="adjustResize" />
 		<activity
 			android:name="yuku.alkitab.base.sync.SyncSettingsActivity"
 			android:label="@string/sync_status_activity_title" />
