+++ Alkitab/src/main/assets/templates/song.css
@@ -1,9 +1,4 @@
 
-body {
-	margin: 0;
-	padding: 24px 16px;
-}
-
 .code {
 	font-family: sans-serif;
 	font-weight: bold;
