+++ Alkitab/src/main/assets/templates/song.html
@@ -21,19 +21,19 @@
 			function body_load() {
 				var test_tobehidden = document.getElementById('test_tobehidden');
 				var test_line = document.getElementById('test_line');
-
+				
 				var h = test_line.clientHeight;
 				var lh = (h * {{$line_spacing_mult}}) + 'px';
-
+				
 				var lines = document.getElementsByClassName('line');
 				for (var i = 0; i < lines.length; i++) lines[i].style.lineHeight = lh;
-
+				
 				test_tobehidden.parentNode.removeChild(test_tobehidden);
 			}
 		</script>
 	</head>
 
-	<body onload="body_load()">
+	<body onload="body_load()" style="padding: 16px 8px">
 		<div class='song'>
 			{{div:code}}
 			{{div:title}}
