+++ Alkitab/src/main/java/yuku/alkitab/base/S.kt
@@ -167,7 +167,6 @@ object S {
         return null // not known
     }
 
-    @JvmStatic
     fun recalculateAppliedValuesBasedOnPreferences() {
         CalculatedDimensionsHolder.applied = calculateDimensionsFromPreferences()
     }
