+++ Alkitab/src/main/java/yuku/alkitab/base/storage/Prefkey.kt
@@ -35,9 +35,6 @@ enum class Prefkey {
     /** Night mode activated (boolean)  */
     is_night_mode,
 
-    /** Night mode follows system theme (boolean)  */
-    follow_system_theme,
-
     /** marker (bookmark) list selected sorting option  */
     marker_list_sort_column, marker_list_sort_ascending,
