+++ Alkitab/src/main/res/layout/activity_devotion_content.xml
@@ -1,42 +1,35 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
 	xmlns:tools="http://schemas.android.com/tools"
-	android:id="@+id/overlayContainer"
 	android:layout_width="match_parent"
-	android:layout_height="match_parent">
+	android:layout_height="match_parent"
+	android:orientation="vertical">
 
-	<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-		xmlns:tools="http://schemas.android.com/tools"
+	<androidx.appcompat.widget.Toolbar
+		android:id="@+id/toolbar"
 		android:layout_width="match_parent"
-		android:layout_height="match_parent"
-		android:orientation="vertical">
+		android:layout_height="?attr/actionBarSize"
+		android:background="?attr/colorPrimary" />
 
-		<androidx.appcompat.widget.Toolbar
-			android:id="@+id/toolbar"
-			android:layout_width="match_parent"
-			android:layout_height="?attr/actionBarSize"
-			android:background="?attr/colorPrimary" />
+	<yuku.alkitab.base.widget.TwofingerLinearLayout
+		android:id="@+id/root"
+		android:layout_width="match_parent"
+		android:layout_height="match_parent">
 
-		<yuku.alkitab.base.widget.TwofingerLinearLayout
-			android:id="@+id/root"
+		<androidx.core.widget.NestedScrollView
+			android:id="@+id/scrollContent"
 			android:layout_width="match_parent"
 			android:layout_height="match_parent">
 
-			<androidx.core.widget.NestedScrollView
-				android:id="@+id/scrollContent"
-				android:layout_width="match_parent"
-				android:layout_height="match_parent">
+			<yuku.alkitab.base.widget.FixedTextView
+				android:id="@+id/lContent"
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content"
+				android:paddingStart="@dimen/text_side_padding"
+				android:paddingEnd="@dimen/text_side_padding"
+				android:textIsSelectable="true"
+				tools:text="Devotion content here." />
+		</androidx.core.widget.NestedScrollView>
 
-				<yuku.alkitab.base.widget.FixedTextView
-					android:id="@+id/lContent"
-					android:layout_width="wrap_content"
-					android:layout_height="wrap_content"
-					android:paddingStart="@dimen/text_side_padding"
-					android:paddingEnd="@dimen/text_side_padding"
-					android:textIsSelectable="true"
-					tools:text="Devotion content here." />
-			</androidx.core.widget.NestedScrollView>
-
-		</yuku.alkitab.base.widget.TwofingerLinearLayout>
-	</LinearLayout>
-</FrameLayout>
+	</yuku.alkitab.base.widget.TwofingerLinearLayout>
+</LinearLayout>
