+++ Alkitab/src/main/res/layout/left_drawer_devotion.xml
@@ -67,11 +67,6 @@
 				android:src="@drawable/ic_nav_end_light" />
 		</LinearLayout>
 
-		<include layout="@layout/left_drawer_display_button" />
-
-		<include
-			layout="@layout/left_drawer_night_mode_switches" />
-
 		<include
 			layout="@layout/left_drawer_footer"
 			android:layout_width="match_parent"
