--- Alkitab/src/main/res/layout/left_drawer_display_button.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Button xmlns:android="http://schemas.android.com/apk/res/android"
-	xmlns:tools="http://schemas.android.com/tools"
-	android:id="@+id/bDisplay"
-	style="?buttonBarButtonStyle"
-	android:layout_width="match_parent"
-	android:layout_height="wrap_content"
-	android:drawablePadding="16dp"
-	android:drawableStart="@drawable/ic_drawer_display"
-	android:fontFamily="sans-serif-regular"
-	android:gravity="start|center_vertical"
-	android:text="@string/menuDisplay"
-	android:textAllCaps="false"
-	android:textColor="#fff"
-	android:textSize="16sp"
-	tools:showIn="@layout/left_drawer_text" />
