--- Alkitab/src/main/res/layout/left_drawer_night_mode_switches.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-	xmlns:tools="http://schemas.android.com/tools"
-	android:layout_width="match_parent"
-	android:layout_height="wrap_content"
-	android:orientation="vertical"
-	android:showDividers="middle"
-	tools:ignore="Overdraw">
-
-	<androidx.appcompat.widget.SwitchCompat
-		android:id="@+id/cNightMode"
-		android:layout_width="match_parent"
-		android:layout_height="48dp"
-		android:layout_marginEnd="16dp"
-		android:layout_marginStart="16dp"
-		android:fontFamily="sans-serif-regular"
-		android:gravity="start|center_vertical"
-		android:text="@string/menuNightMode"
-		android:textColor="#fff"
-		android:textSize="16sp" />
-
-	<androidx.appcompat.widget.SwitchCompat
-		android:id="@+id/cFollowSystemTheme"
-		android:layout_width="match_parent"
-		android:layout_height="48dp"
-		android:layout_marginEnd="16dp"
-		android:layout_marginStart="16dp"
-		android:fontFamily="sans-serif-regular"
-		android:gravity="start|center_vertical"
-		android:text="@string/menuFollowSystemTheme"
-		android:textColor="#fff"
-		android:textSize="16sp" />
-
-</LinearLayout>
