+++ Alkitab/src/main/res/values-it/strings.xml
@@ -366,7 +366,6 @@
     <string name="version_download_network_error">Errore di rete durante il download ^1.</string>
     <string name="version_download_server_error">Errore del server durante il download ^1.</string>
     <string name="menuNightMode">Modalità notturna</string>
-    <string name="menuFollowSystemTheme">Seguire il tema del sistema</string>
     <string name="highlight_custom_color">Altri colori</string>
     <string name="menu_compare">Compara</string>
     <string name="song_player_error_description">Si è verificato un errore durante la riproduzione di questo brano: %1$d, %2$d.</string>
